VenturesLab/requirements.txt
pandas
numpy
plotly
nbformat
sqlalchemy
psycopg2-binary
pysqlite3-binary
matplotlib
watchdog
streamlit-extras